restapi/pyproject.toml
[tool.poetry]
name = "restapi"
version = "0.1.0"
description = ""
authors = [
    "Peder Hovdan Andresen <pedeha@stud.ntnu.no>",
    "Patrick Eilert Krosby <patricek@stud.ntnu.no>",
    "Anders Westby <anderscw@stud.ntnu.no>",
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.73.0"
uvicorn = "^0.17.4"
google-cloud-workflows = "^1.5.0"
loguru = "^0.6.0"
azure-mgmt-logic = "^9.0.0"
firebase-admin = "^5.2.0"
httpx = "^0.22.0"
debugpy = "^1.6.0"
auspex-core = { path = "../core", develop = true }

[tool.poetry.dev-dependencies]
pytest = "7.0.1"
mypy = "^0.931"
black = "^22.1.0"
pytest-cov = "^3.0.0"
hypothesis = "^6.46.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
